truck: move createUserDto to dto folder & update imports
