+++ src/modules/user/interfaces/user.interface.ts
@@ -1,9 +1,3 @@
-import { CreateUserDto } from '../dto/create-user.dto';
-
-export interface CreateUserDtoWithChannel extends CreateUserDto {
-  originChannel: string;
-}
-
 export interface IUser {
   id: string;
   name: string;
